third_party/blink/web_tests/external/wpt/css/css-scroll-anchoring/opt-out-dynamic-expected.txt

This is a testharness.js-based test.
FAIL Dynamically styling 'overflow-anchor: none' on the anchor node should prevent scroll anchoring assert_equals: expected 150 but got 200
Harness: the test ran to completion.
